LPS-193239 Move method to DisplayPageDisplayContext
It's going to be used also in the initial view so it's better to have it in a more generic place
